5.8 alipay

+++ src/main/java/class4/spm/novelnook/pojo/Returned.java
@@ -18,7 +18,7 @@ public class Returned {
     @JsonFormat(pattern = "yyyy-MM-dd")
     Date returntime;
 
-    int fineamount;
+    double fineamount;
 
     boolean ispay;
